Move pytests into tests. Tests 100%
