+++ tests/test_mysqlhandler.py
@@ -4,22 +4,20 @@ Created on 2 Mar 2015
 @author: ph1jb
 
 """
+
+from argparse import Namespace
+from copy import deepcopy
 from datetime import timezone
 from logging import warning
-from os.path import dirname
-from pathlib import Path
-from typing import List, Dict, Any
-from unittest.mock import MagicMock, patch
 import logging
-import os
-import pytest
-import unittest
+from pathlib import Path
+from typing import List
+from unittest.mock import MagicMock, patch
 
 from mysql import connector
 from mysql.connector.errors import (
     DatabaseError,
     DataError,
-    Error,
     IntegrityError,
     InterfaceError,
     InternalError,
@@ -27,77 +25,58 @@ from mysql.connector.errors import (
     OperationalError,
     ProgrammingError,
 )
+import pytest
 import yaml
 
 from mysql_handler import MysqlHandler
 
 
-MYSQL_OPTIONS_DEFAULT = {
-    "autocommit": True,
-    "database": "mysql database not set",
-    "host": "mysql host not set",
-    "password": "mysql password not set",
-    "raise_on_warnings": True,
-    "time_zone": "UTC",
-    "user": "mysql user not set",
-}
-
-secrets_file = Path.joinpath(Path(__file__).parent, "..", "secrets", "test_secrets.yml")
+# pylint: disable=missing-module-docstring
+# pylint: disable=missing-class-docstring
+# pylint: disable=missing-function-docstring
+TABLE = "testtable"
 
 
-@pytest.fixture(autouse=True, scope="module")
-def secrets() -> Dict[str, Any]:
-    """Load secrets file.
-    :return: secrets dict
-    """
-    os.environ["loglevel"] = "DEBUG"
+###########################################
+##Module Fixtures
+###########################################
+@pytest.fixture(scope="module")
+def config_file():
+    return Path(Path(__file__).parent, "test_config.yml")
 
-    with open(secrets_file, "r") as fin:
+
+@pytest.fixture(scope="module")
+def secrets_file():
+    return Path(Path(__file__).parent, "../secrets", "test_secrets.yml")
+
+
+@pytest.fixture(scope="module")
+def secrets_orig(secrets_file):
+    """Read once. Do not update."""
+    with open(secrets_file, "r", encoding="utf8") as fin:
         try:
-            # return yaml.safe_load(fin)
             return yaml.safe_load(fin)
         except yaml.YAMLError as ex:
             warning(ex)
-            raise
+
+
+@pytest.fixture(scope="module")
+def mysql_options_orig(secrets_orig):
+    """Read once. Do not update."""
+    return secrets_orig.get("mysql_options")
+
+
+@pytest.fixture(scope="module")
+def mysql_handler(mysql_options_orig):
+    """Connect to database server, create temporary tables (once) because "if not exist"."""
+    return MysqlHandler(mysql_options_orig)
 
 
 @pytest.fixture(autouse=True, scope="module")
-def mysql_options(secrets) -> Dict[str, Any]:
-    """Load mysql_options.
-    :return: mysql_options
-    """
-    os.environ["loglevel"] = "DEBUG"
-    mysql_options = MYSQL_OPTIONS_DEFAULT
-    mysql_options.update(secrets["mysql_options"])
-    return mysql_options
-
-
-def setUpModule():
-    """Connect to database server, create temporary tables (once) because "if not exist"."""
-    global CURSOR
-    global MYSQL_HANDLER
-    global MYSQL_OPTIONS
-    global ROWS
-    global TABLE
-    MYSQL_OPTIONS = {
-        "autocommit": True,
-        "database": "database_is_not_set",
-        "host": "database_host_is_not_set",
-        "password": "database_password_is_not_set",
-        "raise_on_warnings": True,
-        "time_zone": "UTC",
-        "user": "database_user_is_not_set",
-    }
-    secrets_file = os.path.join(dirname(__file__), "../secrets/", "test_secrets.yml")
-    with open(secrets_file, "r") as fin:
-        try:
-            secrets = yaml.safe_load(fin)
-        except yaml.YAMLError as ex:
-            warning(ex)
-    MYSQL_OPTIONS.update(secrets.get("mysql_options"))
-    cnx = connector.connect(**MYSQL_OPTIONS)
-    CURSOR = cnx.cursor()
-    TABLE = "testtable"
+def db_init(mysql_handler):
+    """Connect to database server, create temporary tables (once) because "if not exist".
+    autouse=True ensures it is called.
+    Scope module ensures it only runs once."""
     statement = f"""CREATE TEMPORARY TABLE if not exists {TABLE} (
                 id int NOT NULL AUTO_INCREMENT,
                 first_name varchar(45) default "AA" NOT NULL,
@@ -105,7 +84,7 @@ def setUpModule():
                  PRIMARY KEY (id)
                 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
                 """
-    CURSOR.execute(statement)
+    mysql_handler.execute(statement)
     statement = f"""CREATE TEMPORARY TABLE if not exists reading30compact (
                 date timestamp NOT NULL DEFAULT '1970-01-02 00:00:00',
                 ss_id int unsigned NOT NULL,
@@ -142,148 +121,188 @@ def setUpModule():
                 PRIMARY KEY (date,ss_id)
                 ) ENGINE=InnoDB ROW_FORMAT=DYNAMIC COMMENT='Prepared Passiv Historic data: SSF ss_id (not Passiv install_id). Omits cols: meter_id,missing_periods, daily_total';
                 """
-    CURSOR.execute(statement)
-    MYSQL_HANDLER = MysqlHandler(mysql_options={}, cnx=cnx)
+    mysql_handler.execute(statement)
 
 
-class Fixture(unittest.TestCase):
-    def setUp(self):
-        global CURSOR
-        global MYSQL_HANDLER
-        global MYSQL_OPTIONS
-        global ROWS
-        global TABLE
-        self.cursor = CURSOR
-        self.mysql_handler = MYSQL_HANDLER
-        self.mysql_options = MYSQL_OPTIONS
-        self.rows = [
-            (
-                1,
-                "Ann",
-                "Awk",
-            ),
-            (2, "Bob", "Bash"),
-            (
-                3,
-                "Cath",
-                "Curl",
-            ),
-            (4, "Dave", "Dig"),
-            (5, "Eve", "Other"),
-        ]
-        self.table = TABLE
-
-    def truncate(self, table: str, foreign_key_checks: int = 1) -> None:
-        """Truncate table (Disable foreign key checks to allow truncate table if foreign keys point to it)"""
-        statement = (
-            f"SET FOREIGN_KEY_CHECKS = {foreign_key_checks};"  # Set foreign key checks (0 disables, 1 enables)
-            f"truncate table {table};"
-            f"SET FOREIGN_KEY_CHECKS = 1;"  # Enable foreign key checks
-        )
-        self.mysql_handler.execute(statement, multi=True)
+###########################################
+##Fixtures (default scope: function)
+###########################################
 
 
-# pylint: disable=missing-module-docstring
-# pylint: disable=missing-class-docstring
-# pylint: disable=missing-function-docstring
+@pytest.fixture()
+def config():
+    return {"loglevel": "INFO"}
 
 
-class TestMysqlHandlerInit(Fixture):
+@pytest.fixture()
+def secrets(secrets_orig):
+    """Refreshed before each use. OK to update."""
+    return deepcopy(secrets_orig)
+
+
+@pytest.fixture()
+def mysql_options_default():
+    """Refreshed before each use. OK to update."""
+    return {
+        "autocommit": True,
+        "database": "database_is_not_set",
+        "host": "database_host_is_not_set",
+        "password": "database_password_is_not_set",
+        "raise_on_warnings": True,
+        "time_zone": "UTC",
+        "user": "database_user_is_not_set",
+    }
+
+
+@pytest.fixture()
+def mysql_options(mysql_options_orig):
+    """Refreshed before each use. OK to update."""
+    return deepcopy(mysql_options_orig)
+
+
+@pytest.fixture()
+def fixture():
+    class Fixture:
+        def __init__(self):
+            self.rows = [
+                (1, "Ann", "Awk"),
+                (2, "Bob", "Bash"),
+                (3, "Cath", "Curl"),
+                (4, "Dave", "Dig"),
+                (5, "Eve", "Other"),
+            ]
+            self.table = TABLE
+            self.cols = ("id", "first_name", "last_name")
+            self.cols_str = ",".join(self.cols)
+
+    return Fixture()
+
+
+@pytest.fixture()
+def truncate(mysql_handler):
+    mysql_handler.execute("truncate test_mysql.testtable")
+
+
+@pytest.fixture()
+def populate(fixture, mysql_handler, truncate):
+    statement = f"insert into {fixture.table} ({fixture.cols_str}) values (%s,%s,%s)"
+    mysql_handler.executemany(statement, fixture.rows)
+
+
+class TestMysqlHandlerInit:
     """Test MysqlHandler.__init__"""
 
-    def setUp(self):
-        super().setUp()
+    def test__init__cnx(self, mysql_options):
+        cnx = connector.connect(**mysql_options)
+        mh = MysqlHandler(mysql_options, cnx=cnx)
+        assert mh.cnx == cnx
+        assert mh.mysql_options == mysql_options
 
-    def test__init__cnx(self):
-        cnx = connector.connect(**self.mysql_options)
-        mh = MysqlHandler(self.mysql_options, cnx=cnx)
-        self.assertEqual(mh.cnx, cnx)
-        self.assertDictEqual(mh.mysql_options, self.mysql_options)
+    def test_with_mysql_handler(self, mysql_options):
+        cnx = connector.connect(**mysql_options)
+        with MysqlHandler(mysql_options, cnx=cnx) as mh:
+            assert isinstance(mh, MysqlHandler)
+            assert mh.cnx == cnx
+            assert mh.mysql_options == mysql_options
 
-    def test_with_mysql_handler(self):
-        with MysqlHandler(self.mysql_options) as mh:
-            self.assertIsInstance(mh, MysqlHandler)
+    @pytest.mark.parametrize(
+        "k,v",
+        [
+            ("database", "mysql database not set"),
+            ("host", "mysql host not set"),
+            ("password", "mysql password not set"),
+            ("user", "mysql user not set"),
+        ],
+    )
+    def test_connection_error(self, k, v, mysql_options):
+        mysql_options.update({k: v})
+        with pytest.raises(connector.errors.Error):
+            MysqlHandler(mysql_options=mysql_options)
 
-    def test_connection_bad_host(self):
-        self.mysql_options = {
-            "database": "mysql database not set",
-            "host": "mysql host not set",
-            "password": "mysql password not set",
-            "user": "mysql user not set",
+    @pytest.mark.parametrize(
+        "arg, key",
+        [
+            ("mysql_database", "database"),
+            ("mysql_host", "host"),
+            ("mysql_password", "password"),
+            ("mysql_user", "user"),
+        ],
+    )
+    def test_override_mysql_options_env(self, arg, key, mysql_options_default) -> None:
+        mysql_args = {
+            "mysql_database": None,
+            "mysql_host": None,
+            "mysql_password": None,
+            "mysql_user": None,
         }
-        self.assertRaises(Error, MysqlHandler, mysql_options=self.mysql_options)
+        mysql_args.update({arg: arg})
+        _config: Namespace = Namespace(
+            **mysql_args, mysql_options=mysql_options_default
+        )
+        MysqlHandler.override_mysql_options(_config)
+        assert _config.mysql_options.get(key) == arg
 
 
-# pylint: disable=missing-function-docstring
-
-
-class TestMysqlHandlerTruncated(Fixture):
+class TestMysqlHandlerTruncated:
     """Test methods needing truncated (empty) tables."""
 
-    def setUp(self):
-        super().setUp()
-        self.truncate(self.table)
-
-    def test_execute(self):
-        statement = f'insert into {self.table} (first_name,last_name) values("A","B")'
-        self.mysql_handler.execute(statement)
+    def test_execute(self, fixture, mysql_handler, truncate):
+        statement = (
+            f'insert into {fixture.table} (first_name,last_name) values("A","B")'
+        )
+        mysql_handler.execute(statement)
         #
         # Check rows have been inserted
-        statement = f"select first_name, last_name from {self.table} limit 1"
-        row = self.mysql_handler.fetchone(statement)
+        statement = f"select first_name, last_name from {fixture.table} limit 1"
+        row = mysql_handler.fetchone(statement)
         expected = ("A", "B")
-        self.assertTupleEqual(row, expected)
+        assert row == expected
 
-    def test_executemany(self):
+    def test_executemany(self, fixture, mysql_handler, truncate):
         statement = (
-            f"insert into {self.table} (id, first_name, last_name) values (%s,%s,%s)"
+            f"insert into {fixture.table} (id, first_name, last_name) values (%s,%s,%s)"
         )
-        self.mysql_handler.executemany(statement, self.rows)
+        mysql_handler.executemany(statement, fixture.rows)
         # Check row3 and row4 have been inserted
-        statement = f"select id, first_name, last_name from {self.table}"
-        rows = self.mysql_handler.fetchall(statement)
-        self.assertListEqual(rows, self.rows)
+        statement = f"select id, first_name, last_name from {fixture.table}"
+        rows = mysql_handler.fetchall(statement)
+        assert rows == fixture.rows
 
-    def test_executemulti(self):
+    def test_executemulti(self, fixture, mysql_handler, truncate):
         statements = (
-            f'insert into {self.table} (first_name,last_name) values("A","B");'
-            f'insert into {self.table} (first_name,last_name) values("C","D")'
+            f'insert into {fixture.table} (first_name,last_name) values("A","B");'
+            f'insert into {fixture.table} (first_name,last_name) values("C","D")'
         )
-        self.mysql_handler.execute(statements, multi=True)
+        mysql_handler.execute(statements, multi=True)
         # Check rows have been inserted
-        statement = f"select first_name, last_name from {self.table}"
-        rows = self.mysql_handler.fetchall(statement)
+        statement = f"select first_name, last_name from {fixture.table}"
+        rows = mysql_handler.fetchall(statement)
         expected = [
             ("A", "B"),
             ("C", "D"),
         ]
-        self.assertListEqual(rows, expected)
+        assert rows == expected
 
 
-class TestMysqlHandlerPopulated(Fixture):
+class TestMysqlHandlerPopulated:
     """Tests which need a populated table."""
 
-    def setUp(self):
-        super().setUp()
-        self.truncate(self.table)
-        # Populate table
-        self.cols = ("id", "first_name", "last_name")
-        self.cols_str = ",".join(self.cols)
-        statement = f"insert into {self.table} ({self.cols_str}) values (%s,%s,%s)"
-        self.cursor.executemany(statement, self.rows)
-
-    def test_insert_on_duplicate_key_update_statement(self):
+    def test_insert_on_duplicate_key_update_statement(
+        self, fixture, mysql_handler, populate
+    ):
         table = "mytable"
         cols = ["a", "b", "c", "d"]
         keys = ["a", "b"]
         on_dup_str = "c=vals.c,d=vals.d"
-        actual = self.mysql_handler.insert_on_duplicate_key_update_statement(
+        actual = mysql_handler.insert_on_duplicate_key_update_statement(
             table, cols, keys
         )
         expected = f"insert into {table} (a,b,c,d) values (%s,%s,%s,%s) as vals on duplicate key update {on_dup_str}"
-        self.assertEqual(actual, expected)
+        assert actual == expected
 
-    def test_insert_select_on_duplicate_key_update_statement(self):
+    def test_insert_select_on_duplicate_key_update_statement(
+        self, fixture, mysql_handler, populate
+    ):
         table_from = "table_from"
         table_into = "table_into"
         colmap = {
@@ -296,7 +315,7 @@ class TestMysqlHandlerPopulated(Fixture):
             "col_into0",
             "col_into1",
         ]
-        actual = self.mysql_handler.insert_select_on_duplicate_key_update_statement(
+        actual = mysql_handler.insert_select_on_duplicate_key_update_statement(
             table_from, table_into, colmap, keys
         )
         expected = (
@@ -305,106 +324,92 @@ class TestMysqlHandlerPopulated(Fixture):
             "as vals(alias0,alias1,alias2,alias3) "
             "on duplicate key update col_into2=vals.alias2,col_into3=vals.alias3"
         )
-        self.assertEqual(actual, expected)
+        assert actual == expected
 
-    def test_fetchall(self):
-        statement = f"select id, first_name, last_name from {self.table} order by id"
-        actual = self.mysql_handler.fetchall(statement)
-        self.assertListEqual(actual, self.rows)
+    def test_fetchall(self, fixture, mysql_handler, populate):
+        statement = f"select id, first_name, last_name from {fixture.table} order by id"
+        actual = mysql_handler.fetchall(statement)
+        assert actual == fixture.rows
 
-    def test_fetchall_data(self):
-        statement = f"select id, first_name, last_name from {self.table} where id in(%s,%s,%s,%s,%s) order by id"
+    def test_fetchall_data(self, fixture, mysql_handler, populate):
+        statement = f"select id, first_name, last_name from {fixture.table} where id in(%s,%s,%s,%s,%s) order by id"
         params = (1, 2, 3, 4, 5)  # [(1,), (2,), (3,), (4,), (5,), ]
-        actual = self.mysql_handler.fetchall(statement, params=params)
-        self.assertListEqual(actual, self.rows)
+        actual = mysql_handler.fetchall(statement, params=params)
+        assert actual == fixture.rows
 
-    def test_fetchone(self):
-        statement = f"select id, first_name, last_name from {self.table} where id = 1  order by id"
-        row = self.mysql_handler.fetchone(statement)
-        self.assertTupleEqual(row, self.rows[0])
+    def test_fetchone(self, fixture, mysql_handler, populate):
+        statement = f"select id, first_name, last_name from {fixture.table} where id = 1  order by id"
+        row = mysql_handler.fetchone(statement)
+        assert row == fixture.rows[0]
 
-    def test_fetchone_param_table_name(self):
+    def test_fetchone_param_table_name(self, fixture, mysql_handler, populate):
         """Cannot parameterise table name. Can only parameterise values."""
         statement = (
             "select id, first_name, last_name from %(table)s where id = 1  order by id"
         )
-        params = {"table": self.table}
-        self.assertRaises(
-            ProgrammingError,
-            self.mysql_handler.fetchone,
-            statement,
-            params=params,
-        )
+        params = {"table": fixture.table}
+        with pytest.raises(connector.errors.ProgrammingError):
+            mysql_handler.fetchone(
+                statement,
+                params=params,
+            )
 
-    def test_fetchone_params(self):
-        statement = f"select id, first_name, last_name from {self.table} where id = %s  order by id"
+    def test_fetchone_params(self, fixture, mysql_handler, populate):
+        statement = f"select id, first_name, last_name from {fixture.table} where id = %s  order by id"
         params = (1,)
-        row = self.mysql_handler.fetchone(statement, params=params)
-        self.assertTupleEqual(row, self.rows[0])
+        row = mysql_handler.fetchone(statement, params=params)
+        assert row == fixture.rows[0]
 
-    def test_fetchone_no_rows_found(self):
+    def test_fetchone_no_rows_found(self, fixture, mysql_handler, populate):
         statement = (
-            f"select id, first_name, last_name from {self.table} where id is null"
+            f"select id, first_name, last_name from {fixture.table} where id is null"
         )
-        row = self.mysql_handler.fetchone(statement)
-        self.assertIsNone(row)
+        row = mysql_handler.fetchone(statement)
+        assert row is None
 
-    def test_insert_on_duplicate_key_update(self):
-        # self.truncate()
-        self.mysql_handler.insert_on_duplicate_key_update(
-            self.table, self.cols, self.cols[:1], self.rows
+    def test_insert_on_duplicate_key_update(self, fixture, mysql_handler, populate):
+        mysql_handler.insert_on_duplicate_key_update(
+            fixture.table, fixture.cols, fixture.cols[:1], fixture.rows
         )
         # Re-insert to show that on duplicate key update works
-        self.mysql_handler.insert_on_duplicate_key_update(
-            self.table, self.cols, self.cols[:1], self.rows
+        mysql_handler.insert_on_duplicate_key_update(
+            fixture.table, fixture.cols, fixture.cols[:1], fixture.rows
         )
         # Check rows have been inserted
-        statement = f"select * from {self.table}"
-        rows = self.mysql_handler.fetchall(statement)
-        self.assertListEqual(rows, self.rows)
+        statement = f"select * from {fixture.table}"
+        rows = mysql_handler.fetchall(statement)
+        assert rows == fixture.rows
 
-    def test_insert_on_duplicate_key_update_on_dup(self):
-        # self.truncate()
+    def test_insert_on_duplicate_key_update_on_dup(
+        self, fixture, mysql_handler, populate
+    ):
         on_dup = "first_name=vals.first_name,last_name=UPPER(vals.last_name)"
-        self.mysql_handler.insert_on_duplicate_key_update(
-            self.table, self.cols, self.cols[:1], self.rows, on_dup=on_dup
+        mysql_handler.insert_on_duplicate_key_update(
+            fixture.table, fixture.cols, fixture.cols[:1], fixture.rows, on_dup=on_dup
         )
         # Re-insert to show that on duplicate key update works
-        self.mysql_handler.insert_on_duplicate_key_update(
-            self.table, self.cols, self.cols[:1], self.rows, on_dup=on_dup
+        mysql_handler.insert_on_duplicate_key_update(
+            fixture.table, fixture.cols, fixture.cols[:1], fixture.rows, on_dup=on_dup
         )
         # Check rows have been inserted
-        statement = f"select * from {self.table}"
-        rows = self.mysql_handler.fetchall(statement)
+        statement = f"select * from {fixture.table}"
+        rows = mysql_handler.fetchall(statement)
         rows_uc = [
-            (
-                1,
-                "Ann",
-                "AWK",
-            ),
+            (1, "Ann", "AWK"),
             (2, "Bob", "BASH"),
-            (
-                3,
-                "Cath",
-                "CURL",
-            ),
+            (3, "Cath", "CURL"),
             (4, "Dave", "DIG"),
             (5, "Eve", "OTHER"),
         ]
+        assert rows == rows_uc
 
-        self.assertListEqual(rows, rows_uc)
-
-    def test_on_dup(self):
-        col_names = [
-            "a",
-            "b",
-            "c",
-        ]
-        actual = self.mysql_handler.on_dup(col_names)
+    def test_on_dup(self, fixture, mysql_handler, populate):
+        col_names = ["a", "b", "c"]
+        actual = mysql_handler.on_dup(col_names)
         expected = "a=vals.a,b=vals.b,c=vals.c"
-        self.assertEqual(actual, expected)
+        assert actual == expected
 
-    def test_close_cursor(self):
+    def test_close_cursor(self, fixture, mysql_handler, populate):
         """
         https://dev.mysql.com/doc/connector-python/en/connector-python-api-errors-error.html
         # Error handling
@@ -412,26 +417,21 @@ class TestMysqlHandlerPopulated(Fixture):
         NotSupportedError, OperationalError, ProgrammingError, Warning
         """
         with patch.object(
-            self.mysql_handler.cnx, "cursor"
+            mysql_handler.cnx, "cursor"
         ) as cursor_fn:  # pylint: disable:undefined-variable
             cursor = cursor_fn()
-            cursor.fetchall = MagicMock(return_value=self.rows)
-            statement = f"select * from {self.table}"
-            actual = self.mysql_handler.fetchall(statement)
-            self.assertListEqual(actual, self.rows)
+            cursor.fetchall = MagicMock(return_value=fixture.rows)
+            statement = f"select * from {fixture.table}"
+            actual = mysql_handler.fetchall(statement)
+            assert actual == fixture.rows
             cursor.close.assert_called_once_with()
 
-    def test_close_cursor_on_exception(self):
-        """
-        https://dev.mysql.com/doc/connector-python/en/connector-python-api-errors-error.html
-        # Error handling
-        DataError, DatabaseError, Error, IntegrityError, InterfaceError, InternalError,
-        NotSupportedError, OperationalError, ProgrammingError, Warning
-        """
-        for ex in (
+    @pytest.mark.parametrize(
+        "ex",
+        [
             DataError,
             DatabaseError,
-            Error,
+            connector.errors.Error,
             IntegrityError,
             InterfaceError,
             InternalError,
@@ -439,37 +439,34 @@ class TestMysqlHandlerPopulated(Fixture):
             OperationalError,
             ProgrammingError,
             Warning,
-        ):
-            with self.subTest():
-                with patch.object(
-                    self.mysql_handler.cnx, "cursor"
-                ) as cursor_fn:  # pylint: disable:undefined-variable
-                    cursor = cursor_fn()
-                    cursor.execute = MagicMock(side_effect=ex())
-                    statement = "select * from XXX"  # non existent table
-                    self.assertRaises(ex, self.mysql_handler.execute, statement)
-                    cursor.close.assert_called_once_with()
+        ],
+    )
+    def test_close_cursor_on_exception(self, ex, fixture, mysql_handler, populate):
+        """
+        https://dev.mysql.com/doc/connector-python/en/connector-python-api-errors-error.html
+        # Error handling
+        DataError, DatabaseError, Error, IntegrityError, InterfaceError, InternalError,
+        NotSupportedError, OperationalError, ProgrammingError, Warning
+        """
+        with patch.object(
+            mysql_handler.cnx, "cursor"
+        ) as cursor_fn:  # pylint: disable:undefined-variable
+            cursor = cursor_fn()
+            cursor.execute = MagicMock(side_effect=ex())
+            statement = "select * from XXX"  # non existent table
+            with pytest.raises(ex):
+                mysql_handler.execute(statement)
+            cursor.close.assert_called_once_with()
 
     # def test_reset_auto_increment(self):
     #     table = 'testtable'
     #     col = 'id'
     #     statement = f'select max({col}) from {table}'
-    #     max_val = self.mysql_handler.fetchone(statement)[0]
-    #     auto_increment = self.mysql_handler.reset_auto_increment(table, col)
+    #     max_val = mysql_handler.fetchone(statement)[0]
+    #     auto_increment = mysql_handler.reset_auto_increment(table, col)
     #     self.assertEqual(auto_increment, max_val + 1)
 
-    def test_type_conf(self):
-        def lister(list_arg: List[str]):
-            return list_arg
-
-        list_arg = lister(0)
-        print(type(list_arg))
-        print(list_arg)
-        self.assertEqual(list_arg, 0)
-
-    def test_insert_on_duplicate_key_long(self):
-        # n = 6
-        # n = 24
+    def test_insert_on_duplicate_key_long(self, fixture, mysql_handler, populate):
         n = 26
         table = "reading30compact"
         keys = ["date", "ss_id"]
@@ -478,10 +475,9 @@ class TestMysqlHandlerPopulated(Fixture):
         date = "2022-01-01"
         ss_id = 1
         rows = [(date, ss_id, *vals)]
-        self.mysql_handler.insert_on_duplicate_key_update(table, cols, keys, rows)
-        # print(self.mysql_handler.execute('show tables'))
+        mysql_handler.insert_on_duplicate_key_update(table, cols, keys, rows)
 
-    def test_timestamp_types(self):
+    def test_timestamp_types(self, fixture, mysql_handler, populate):
         """
         Determining if an Object is Aware or Naive
         Objects of the date type are always naive.
@@ -497,13 +493,13 @@ class TestMysqlHandlerPopulated(Fixture):
         The distinction between aware and naive doesn’t apply to timedelta objects.
         """
         statement = "select date from reading30compact"
-        (timestamp,) = self.mysql_handler.fetchone(statement)
+        (timestamp,) = mysql_handler.fetchone(statement)
 
         print(timestamp, type(timestamp), timestamp.tzinfo)
         timestamp_aware = timestamp.astimezone(timezone.utc)
         print(timestamp_aware, type(timestamp_aware), timestamp_aware.tzinfo)
 
-        (timestamp2,) = self.mysql_handler.fetchone(statement)
+        (timestamp2,) = mysql_handler.fetchone(statement)
         print(timestamp2, type(timestamp2), timestamp2.tzinfo)
 
 
@@ -515,13 +511,13 @@ class TestMysqlExceptionHandling:
         caplog.set_level(logging.INFO)
         msg = "hi"
         errno = 123
-        error = Error(msg, 123)
-        with pytest.raises(Error, match=msg) as cm_ex:
+        error = connector.Error(msg, 123)
+        with pytest.raises(connector.Error, match=msg) as cm_ex:
             with MysqlHandler(mysql_options):
-                raise Error(msg=msg, errno=errno)
+                raise connector.Error(msg=msg, errno=errno)
             # Exception
-            self.assertEqual(cm_ex.type, Error)
-            self.assertEqual(str(cm_ex.value), str(error))
+            assert cm_ex.type == connector.Error
+            assert str(cm_ex.value) == str(error)
         # Logging
         for record in caplog.records:
             assert record.levelname == "CRITICAL"
@@ -529,12 +525,12 @@ class TestMysqlExceptionHandling:
 
     def test_execute_exception(self, caplog, mysql_options):
         caplog.set_level(logging.INFO)
-        with pytest.raises(Error) as cm_ex:
+        with pytest.raises(connector.Error) as cm_ex:
             with MysqlHandler(mysql_options) as mh:
                 statement = "select no_col from no_table"
                 mh.execute(statement)
             # Exception
-            assert cm_ex.type == Error
+            assert cm_ex.type == connector.Error
         # Logging
         for record in caplog.records:
             assert record.levelname == "CRITICAL"
@@ -542,12 +538,12 @@ class TestMysqlExceptionHandling:
 
     def test_fetchone_exception(self, caplog, mysql_options):
         caplog.set_level(logging.INFO)
-        with pytest.raises(Error) as cm_ex:
+        with pytest.raises(connector.Error) as cm_ex:
             with MysqlHandler(mysql_options) as mh:
                 statement = "select no_col from no_table"
                 mh.fetchone(statement)
             # Exception
-            assert cm_ex.type == Error
+            assert cm_ex.type == connector.Error
         # Logging
         for record in caplog.records:
             assert record.levelname == "CRITICAL"
@@ -555,12 +551,12 @@ class TestMysqlExceptionHandling:
 
     def test_fetchall_exception(self, caplog, mysql_options):
         caplog.set_level(logging.INFO)
-        with pytest.raises(Error) as cm_ex:
+        with pytest.raises(connector.Error) as cm_ex:
             with MysqlHandler(mysql_options) as mh:
                 statement = "select no_col from no_table"
                 mh.fetchall(statement)
             # Exception
-            assert cm_ex.type == Error
+            assert cm_ex.type == connector.Error
         # Logging
         for record in caplog.records:
             assert record.levelname == "CRITICAL"
@@ -582,14 +578,10 @@ class TestMysqlExceptionHandling:
             "col_into0",
             "col_into1",
         ]
-        with pytest.raises(Error) as cm_ex, MysqlHandler(mysql_options) as mh:
+        with pytest.raises(connector.Error) as cm_ex, MysqlHandler(mysql_options) as mh:
             mh.insert_on_duplicate_key_update(table_from, table_into, colmap, keys)
         # Exception
-        assert cm_ex.type == InterfaceError
+        assert cm_ex.type == connector.errors.InterfaceError
         # Logging
         for record in caplog.records:
             assert record.levelname == "CRITICAL"
-
-
-if __name__ == "__main__":
-    unittest.main()
